Extracted class RunningExecutionsUpdater from RunningExecutionsUpdaterThread
+ Added copyright & javadocs to both
